new
new

+++ app/Http/Controllers/Api/EmployerController.php
@@ -63,12 +63,12 @@ class EmployerController extends Controller
             $request->company_name ? $profile->company_name = $request->company_name : '';
             $request->language ? $profile->language = implode(',', $request->language) : '';
             if ($request->hasfile('image')) {
-                if (!empty($profile->image)) {
+                if (!empty($profile->image) && ($profile->image != "assets/dist/img/avatar5.png")) {
                     $image_path = $profile->image;
                     unlink($image_path);
                 }
                 $image = $request->file('image');
-                $name = time() . 'profile' . '.' . $image->getClientOriginalExtension();
+                $name = $id . 'profile' . '.' . $image->getClientOriginalExtension();
                 $destinationPath = 'profile_images/';
                 $image->move($destinationPath, $name);
                 $profile->image = 'profile_images/' . $name;
@@ -81,7 +81,7 @@ class EmployerController extends Controller
                     unlink($logo_path);
                 }
                 $company_logo = $request->file('company_logo');
-                $name = $id . 'company_logo' . '.' . $company_logo->getClientOriginalExtension();
+                $name = time() . 'company_logo' . '.' . $company_logo->getClientOriginalExtension();
                 $destinationPath = 'company_logo/';
                 $company_logo->move($destinationPath, $name);
                 $profile->company_logo = 'company_logo/' . $name;
@@ -93,7 +93,7 @@ class EmployerController extends Controller
                     unlink($video_path);
                 }
                 $video = $request->file('video');
-                $name = $id . 'profile_video' . '.' . $video->getClientOriginalExtension();
+                $name = time() . 'profile_video' . '.' . $video->getClientOriginalExtension();
                 $destinationPath = 'profile_videos/';
                 $video->move($destinationPath, $name);
                 $profile->video = 'profile_videos/' . $name;
